Property router, service, controller
