+++ src/controllers/userController.js
@@ -79,42 +79,11 @@ export class UserController {
     }
   };
 
-  allProperties = async (req, res) => {
+  userById = async (req, res) => {
     try {
-      // Desestruturação dos parâmetros da query diretamente do req.query
-      const {
-        page = 1,
-        limit = 10,
-        sort = "price",
-        type,
-        search,
-        bedrooms,
-        bathrooms,
-        minPrice,
-        maxPrice,
-        minSize,
-        maxSize,
-        features,
-      } = req.query;
-
-      // Passar os parâmetros para a função de serviço
-      const data = {
-        page,
-        limit,
-        sort,
-        type,
-        search,
-        bedrooms,
-        bathrooms,
-        minPrice,
-        maxPrice,
-        minSize,
-        maxSize,
-        features,
-      };
-      // Chamada do serviço com os dados processados
-      const response = await this.userServices.allProperties(data);
-      res.status(HttpStatus.OK).json({ response });
+      const data = req.params.id;
+      const response = await this.userServices.userById(data);
+      res.status(HttpStatus.OK).json(response);
     } catch (error) {
       res
         .status(HttpStatus.INTERNAL_SERVER_ERROR)
