+++ src/models/usersModel.js
@@ -12,26 +12,21 @@ const roleHierarchy = {
   [roles.ADMIN]: 3,
 };
 
-const UserSchema = new mongoose.Schema(
-  {
-    name: { type: String, required: true }, //User
-    dateOfBirth: { type: Date, required: true }, //User
-    email: { type: String, required: true, unique: true }, //User
-    password: { type: String, required: true }, //User
-    role: {
-      type: String,
-      enum: Object.values(roles),
-      required: true,
-      default: roles.CLIENT,
-    }, //User
-    resetPasswordToken: { type: String }, //User
-    phone: { type: String }, //User ?
-    agentLicense: { type: String }, //Agent
-    employer: { type: String }, //Agent
-    properties: [{ type: mongoose.Schema.Types.ObjectId, ref: "Property" }], //Agent
-  },
-  { discriminatorKey: "role" }
-);
+const UserSchema = new mongoose.Schema({
+  name: { type: String, required: true }, //User
+  email: { type: String, required: true, unique: true }, //User
+  password: { type: String, required: true }, //User
+  role: {
+    type: String,
+    enum: Object.values(roles),
+    required: true,
+    default: roles.CLIENT,
+  }, //User
+  resetPasswordToken: { type: String }, //User
+  phone: { type: String }, //User ?
+  agentLicense: { type: String }, //Agent
+  employer: { type: String }, //Agent onde trabalha
+});
 
 const User = mongoose.model("User", UserSchema);
